sort map autocomplete options
